requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
cssutils==1.0.1
cycler==0.10.0
httmock==1.2.5
inflect==0.2.5
jaraco.itertools==1.7.1
kombu==3.0.35
lxml==3.6.0
more-itertools==2.2
numpy==1.11.0
pyparsing==2.1.5
python-dateutil==2.5.3
pytz==2016.4
redis==2.10.5
requests==2.10.0
six==1.10.0
svg.charts==3.3
tempora==1.4
